feat: refactor mobile detection hook; rename use-mobile to use-is-mobile and update imports
