+++ src/hooks/use-is-mobile.ts
@@ -1,3 +1,5 @@
+'use client'
+
 import { useEffect, useState } from 'react'
 
 const MOBILE_BREAKPOINT = 768
